# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Esaal medical appointment platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /doctor/slots:
    get:
      operationId: getDoctorOwnSlots
      tags: [slots]
      summary: Get current doctor's slots
      responses:
        "200":
          description: Doctor's slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Slot"
    post:
      operationId: createSlot
      tags: [slots]
      summary: Create an availability slot (doctor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSlotBody"
      responses:
        "201":
          description: Slot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Slot"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /doctor/slots/{id}:
    delete:
      operationId: deleteSlot
      tags: [slots]
      summary: Delete a slot (doctor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Slot deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Slot:
      type: object
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        isBooked:
          type: boolean
      required:
        - id
        - doctorId
        - startTime
        - endTime
        - isBooked
    CreateSlotBody:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
      required:
        - startTime
        - endTime
...